chat bot using gemini api
