+++ front-end/my-vite-app/.env
@@ -1,3 +1,4 @@
 VITE_BACKEND_ADDRESS = http://localhost:3000
 VITE_PHOTO_API_KEY = rQAayG_V9eZA2-BB9-4sGYJajQP6og3W8EbaIC1dd_g
 VITE_YOUTUBE_API_KEY = AIzaSyBUgrG3C2rHVN6rdCHikXuzboRyijMAZ0o
+VITE_GEMINI_API_KEY = AIzaSyCN-IDraq7tWuUS6zTrDgyh5dMs5wyk6rs
