+++ front-end/my-vite-app/src/CoursePage.css
@@ -24,7 +24,7 @@
 }
 
 .course{
-    width: 30vh;
+    width: 35vh;
     background-color:#ffffff;
     padding: 15px;
     box-shadow:0 0 40px -10px rgba(0,0,0,.3), 0 0 25px -15px rgba(0,0,0,.2);
@@ -38,9 +38,8 @@
 }
 
 .course-img{
-    width: 30vh;
+    width: 35vh;
     height: 30vh;
-    padding: 5px;
 }
 
 .coursepage{
